fix(schema): Allow GenericOf(Object) usage
