+++ packages/schema/src/utils/getJsonSchema.spec.ts
@@ -1,20 +1,7 @@
-import {ancestorsOf, nameOf} from "@tsed/core";
-import {Type} from "@tsed/core/src/interfaces/Type";
+import {ancestorsOf, nameOf, Type} from "@tsed/core";
 import {expect} from "chai";
 import {Post} from "../../test/helpers/Post";
-import {
-  CollectionOf,
-  Email,
-  Format,
-  GenericOf,
-  Generics,
-  getJsonSchema,
-  JsonEntityStore,
-  MinLength,
-  Name,
-  Property,
-  Required
-} from "../index";
+import {CollectionOf, Email, Format, getJsonSchema, JsonEntityStore, MinLength, Name, Property, Required} from "../index";
 
 describe("getJsonSchema", () => {
   it("should declare all schema correctly (basic)", () => {
@@ -628,142 +615,6 @@ describe("getJsonSchema", () => {
         }
       },
       required: ["email"]
-    });
-  });
-  it("should return the json schema for an inherited model and generics", () => {
-    @Generics("T")
-    class Base<T> {
-      @Property()
-      id: string;
-
-      @Required()
-      @Email()
-      email: string;
-
-      @Property("T")
-      role: T;
-    }
-
-    class Model<T> extends Base<T> {
-      @MinLength(0)
-      email: string;
-
-      @Property()
-      name: string;
-    }
-
-    class Role {
-      @Property()
-      level: string;
-    }
-
-    class Content {
-      @GenericOf(Role)
-      payload: Model<Role>;
-    }
-
-    expect(getJsonSchema(Content)).to.deep.eq({
-      definitions: {
-        Role: {
-          properties: {
-            level: {
-              type: "string"
-            }
-          },
-          type: "object"
-        }
-      },
-      properties: {
-        payload: {
-          properties: {
-            email: {
-              format: "email",
-              minLength: 0,
-              type: "string"
-            },
-            id: {
-              type: "string"
-            },
-            name: {
-              type: "string"
-            },
-            role: {
-              $ref: "#/definitions/Role"
-            }
-          },
-          required: ["email"],
-          type: "object"
-        }
-      },
-      type: "object"
-    });
-  });
-  it("should return the json schema with hosted schemes", () => {
-    @Generics("T")
-    class Base<T> {
-      @Property()
-      id: string;
-
-      @Required()
-      @Email()
-      email: string;
-
-      @Property("T")
-      role: T;
-    }
-
-    class Model<T> extends Base<T> {
-      @MinLength(0)
-      email: string;
-
-      @Property()
-      name: string;
-    }
-
-    class Role {
-      @Property()
-      level: string;
-    }
-
-    class Content {
-      @GenericOf(Role)
-      payload: Model<Role>;
-    }
-
-    expect(getJsonSchema(Content, {host: "http://example.com/schema"})).to.deep.eq({
-      definitions: {
-        Role: {
-          properties: {
-            level: {
-              type: "string"
-            }
-          },
-          type: "object"
-        }
-      },
-      properties: {
-        payload: {
-          properties: {
-            email: {
-              format: "email",
-              minLength: 0,
-              type: "string"
-            },
-            id: {
-              type: "string"
-            },
-            name: {
-              type: "string"
-            },
-            role: {
-              $ref: "http://example.com/schema/Role"
-            }
-          },
-          required: ["email"],
-          type: "object"
-        }
-      },
-      type: "object"
     });
   });
   it("should keep required meta from inherited class", () => {
