+++ packages/schema/src/utils/getSpec.spec.ts
@@ -1,7 +1,7 @@
 import {expect} from "chai";
 import {validateSpec} from "../../test/helpers/validateSpec";
 import {Consumes, In, Min, Name, OperationPath, Required} from "../decorators";
-import {CollectionOf, Description, GenericOf, Generics, getJsonSchema, Property, Returns, SpecTypes} from "../index";
+import {CollectionOf, Description, Property, Returns, SpecTypes} from "../index";
 import {getSpec} from "./getSpec";
 
 describe("getSpec()", () => {
@@ -1021,382 +1021,6 @@ describe("getSpec()", () => {
           }
         });
       });
-      it("should declare all schema correctly (generics - openspec2)", () => {
-        // WHEN
-        @Generics("T")
-        class Submission<T> {
-          @Property()
-          _id: string;
-
-          @Property("T")
-          data: T;
-        }
-
-        class Product {
-          @Property()
-          title: string;
-        }
-
-        class Article {
-          @Property()
-          id: string;
-        }
-
-        class Controller1 {
-          @OperationPath("POST", "/")
-          async method(@In("body") @GenericOf(Product) submission: Submission<Product>) {
-            return null;
-          }
-        }
-
-        class Controller2 {
-          @OperationPath("POST", "/")
-          async method(@In("body") @GenericOf(Article) submission: Submission<Article>) {
-            return null;
-          }
-        }
-
-        // THEN
-        const spec1 = getSpec(Controller1, {specType: SpecTypes.SWAGGER});
-        const spec2 = getSpec(Controller2, {specType: SpecTypes.SWAGGER});
-
-        expect(spec1).to.deep.equal({
-          definitions: {
-            Product: {
-              properties: {
-                title: {
-                  type: "string"
-                }
-              },
-              type: "object"
-            },
-            Submission: {
-              properties: {
-                _id: {
-                  type: "string"
-                },
-                data: {
-                  $ref: "#/definitions/Product"
-                }
-              },
-              type: "object"
-            }
-          },
-          tags: [
-            {
-              name: "Controller1"
-            }
-          ],
-          paths: {
-            "/": {
-              post: {
-                operationId: "controller1Method",
-                tags: ["Controller1"],
-                parameters: [
-                  {
-                    in: "body",
-                    name: "body",
-                    required: false,
-                    schema: {
-                      $ref: "#/definitions/Submission"
-                    }
-                  }
-                ],
-                responses: {
-                  "200": {
-                    description: "Success"
-                  }
-                }
-              }
-            }
-          }
-        });
-        expect(spec2).to.deep.equal({
-          definitions: {
-            Article: {
-              properties: {
-                id: {
-                  type: "string"
-                }
-              },
-              type: "object"
-            },
-            Submission: {
-              properties: {
-                _id: {
-                  type: "string"
-                },
-                data: {
-                  $ref: "#/definitions/Article"
-                }
-              },
-              type: "object"
-            }
-          },
-          tags: [
-            {
-              name: "Controller2"
-            }
-          ],
-          paths: {
-            "/": {
-              post: {
-                operationId: "controller2Method",
-                tags: ["Controller2"],
-                parameters: [
-                  {
-                    in: "body",
-                    name: "body",
-                    required: false,
-                    schema: {
-                      $ref: "#/definitions/Submission"
-                    }
-                  }
-                ],
-                responses: {
-                  "200": {
-                    description: "Success"
-                  }
-                }
-              }
-            }
-          }
-        });
-      });
-      it("should declare all schema correctly (generics - openspec3)", () => {
-        // WHEN
-        @Generics("T")
-        class Submission<T> {
-          @Property()
-          _id: string;
-
-          @Property("T")
-          data: T;
-        }
-
-        class Product {
-          @Property()
-          title: string;
-        }
-
-        class Controller1 {
-          @OperationPath("POST", "/")
-          async method(@In("body") @GenericOf(Product) submission: Submission<Product>) {
-            return null;
-          }
-        }
-
-        // THEN
-        const spec1 = getSpec(Controller1, {specType: SpecTypes.OPENAPI});
-
-        expect(spec1).to.deep.equal({
-          components: {
-            schemas: {
-              Product: {
-                properties: {
-                  title: {
-                    type: "string"
-                  }
-                },
-                type: "object"
-              },
-              Submission: {
-                properties: {
-                  _id: {
-                    type: "string"
-                  },
-                  data: {
-                    $ref: "#/components/schemas/Product"
-                  }
-                },
-                type: "object"
-              }
-            }
-          },
-          paths: {
-            "/": {
-              post: {
-                operationId: "controller1Method",
-                parameters: [],
-                requestBody: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        $ref: "#/components/schemas/Submission"
-                      }
-                    }
-                  },
-                  required: false
-                },
-                responses: {
-                  "200": {
-                    description: "Success"
-                  }
-                },
-                tags: ["Controller1"]
-              }
-            }
-          },
-          tags: [
-            {
-              name: "Controller1"
-            }
-          ]
-        });
-      });
-      it("should declare all schema correctly (generics - nested - openspec3 and 2)", () => {
-        // WHEN
-        @Generics("T")
-        class Pagination<T> {
-          @CollectionOf("T")
-          data: T[];
-
-          @Property()
-          totalCount: number;
-        }
-
-        @Generics("T")
-        class Submission<T> {
-          @Property()
-          _id: string;
-
-          @Property("T")
-          data: T;
-        }
-
-        class Product {
-          @Property()
-          title: string;
-        }
-
-        class MyController {
-          @OperationPath("POST", "/")
-          @(Returns(200, Pagination).Of(Submission).Nested(Product).Description("description"))
-          async method(): Promise<Pagination<Submission<Product>> | null> {
-            return null;
-          }
-        }
-
-        // THEN
-        const spec1 = getSpec(MyController, {specType: SpecTypes.OPENAPI});
-
-        expect(spec1).to.deep.equal({
-          components: {
-            schemas: {
-              Product: {
-                properties: {
-                  title: {
-                    type: "string"
-                  }
-                },
-                type: "object"
-              }
-            }
-          },
-          paths: {
-            "/": {
-              post: {
-                operationId: "myControllerMethod",
-                parameters: [],
-                responses: {
-                  "200": {
-                    content: {
-                      "application/json": {
-                        schema: {
-                          properties: {
-                            data: {
-                              items: {
-                                properties: {
-                                  _id: {
-                                    type: "string"
-                                  },
-                                  data: {
-                                    $ref: "#/components/schemas/Product"
-                                  }
-                                },
-                                type: "object"
-                              },
-                              type: "array"
-                            },
-                            totalCount: {
-                              type: "number"
-                            }
-                          },
-                          type: "object"
-                        }
-                      }
-                    },
-                    description: "description"
-                  }
-                },
-                tags: ["MyController"]
-              }
-            }
-          },
-          tags: [
-            {
-              name: "MyController"
-            }
-          ]
-        });
-
-        const spec2 = getSpec(MyController, {specType: SpecTypes.SWAGGER});
-
-        expect(spec2).to.deep.equal({
-          definitions: {
-            Product: {
-              properties: {
-                title: {
-                  type: "string"
-                }
-              },
-              type: "object"
-            }
-          },
-          paths: {
-            "/": {
-              post: {
-                operationId: "myControllerMethod",
-                parameters: [],
-                produces: ["application/json"],
-                responses: {
-                  "200": {
-                    description: "description",
-                    schema: {
-                      properties: {
-                        data: {
-                          items: {
-                            properties: {
-                              _id: {
-                                type: "string"
-                              },
-                              data: {
-                                $ref: "#/definitions/Product"
-                              }
-                            },
-                            type: "object"
-                          },
-                          type: "array"
-                        },
-                        totalCount: {
-                          type: "number"
-                        }
-                      },
-                      type: "object"
-                    }
-                  }
-                },
-                tags: ["MyController"]
-              }
-            }
-          },
-          tags: [
-            {
-              name: "MyController"
-            }
-          ]
-        });
-      });
     });
   });
 
@@ -1552,440 +1176,6 @@ describe("getSpec()", () => {
                           type: "string"
                         },
                         type: "array"
-                      }
-                    }
-                  },
-                  description: "description"
-                }
-              }
-            }
-          }
-        }
-      });
-    });
-    it("should declare an Generic of Model (swagger2)", async () => {
-      // WHEN
-      @Generics("T")
-      class Pagination<T> {
-        @CollectionOf("T")
-        data: T[];
-
-        @Property()
-        totalCount: number;
-      }
-
-      class Product {
-        @Property()
-        title: string;
-      }
-
-      class Controller {
-        @OperationPath("POST", "/")
-        @(Returns(200, Pagination).Of(Product).Description("description"))
-        async method(): Promise<Pagination<Product> | null> {
-          return null;
-        }
-      }
-
-      // THEN
-      const spec = getSpec(Controller, {specType: SpecTypes.SWAGGER});
-
-      expect(spec).to.deep.equal({
-        definitions: {
-          Product: {
-            properties: {
-              title: {
-                type: "string"
-              }
-            },
-            type: "object"
-          }
-        },
-        tags: [
-          {
-            name: "Controller"
-          }
-        ],
-        paths: {
-          "/": {
-            post: {
-              operationId: "controllerMethod",
-              parameters: [],
-              produces: ["application/json"],
-              tags: ["Controller"],
-              responses: {
-                "200": {
-                  description: "description",
-                  schema: {
-                    properties: {
-                      data: {
-                        items: {
-                          $ref: "#/definitions/Product"
-                        },
-                        type: "array"
-                      },
-                      totalCount: {
-                        type: "number"
-                      }
-                    },
-                    type: "object"
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-    });
-    it("should declare an Generic of Model (openspec3)", async () => {
-      // WHEN
-      @Generics("T")
-      class Pagination<T> {
-        @CollectionOf("T")
-        data: T[];
-
-        @Property()
-        totalCount: number;
-      }
-
-      @Generics("T")
-      class Submission<T> {
-        @Property()
-        _id: string;
-
-        @Property("T")
-        data: T;
-      }
-
-      class Product {
-        @Property()
-        title: string;
-      }
-
-      class Controller {
-        @OperationPath("POST", "/")
-        @(Returns(200, Pagination).Of(Submission).Nested(Product).Description("description"))
-        async method(): Promise<Pagination<Submission<Product>> | null> {
-          return null;
-        }
-      }
-
-      expect(getJsonSchema(Pagination)).to.deep.eq({
-        properties: {
-          data: {
-            items: {
-              $ref: "T"
-            },
-            type: "array"
-          },
-          totalCount: {
-            type: "number"
-          }
-        },
-        type: "object"
-      });
-
-      // THEN
-      const spec = getSpec(Controller, {specType: SpecTypes.OPENAPI});
-
-      expect(spec).to.deep.equal({
-        components: {
-          schemas: {
-            Product: {
-              properties: {
-                title: {
-                  type: "string"
-                }
-              },
-              type: "object"
-            }
-          }
-        },
-        tags: [
-          {
-            name: "Controller"
-          }
-        ],
-        paths: {
-          "/": {
-            post: {
-              operationId: "controllerMethod",
-              parameters: [],
-              tags: ["Controller"],
-              responses: {
-                "200": {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        properties: {
-                          data: {
-                            items: {
-                              properties: {
-                                data: {
-                                  $ref: "#/components/schemas/Product"
-                                },
-                                _id: {
-                                  type: "string"
-                                }
-                              },
-                              type: "object"
-                            },
-                            type: "array"
-                          },
-                          totalCount: {
-                            type: "number"
-                          }
-                        },
-                        type: "object"
-                      }
-                    }
-                  },
-                  description: "description"
-                }
-              }
-            }
-          }
-        }
-      });
-    });
-    it("should declare a nested Generics of Model (swagger2)", () => {
-      // WHEN
-      @Generics("T")
-      class Pagination<T> {
-        @CollectionOf("T")
-        data: string;
-
-        @Property()
-        totalCount: number;
-      }
-
-      @Generics("T")
-      class Submission<T> {
-        @Property()
-        _id: string;
-
-        @Property("T")
-        data: T;
-      }
-
-      class Product {
-        @Property()
-        title: string;
-      }
-
-      class Article {
-        @Property()
-        id: string;
-      }
-
-      class Controller {
-        @OperationPath("POST", "/")
-        @(Returns(200, Pagination).Of(Submission).Nested(Product).Description("description"))
-        method(): Pagination<Submission<Product>> | null {
-          return null;
-        }
-      }
-
-      class Controller2 {
-        @OperationPath("POST", "/")
-        @(Returns(200, Pagination).Of(Submission).Nested(Article).Description("description"))
-        method(): Pagination<Submission<Article>> | null {
-          return null;
-        }
-      }
-
-      // THEN
-      const spec2 = getSpec(Controller2, {specType: SpecTypes.SWAGGER});
-      const spec = getSpec(Controller, {specType: SpecTypes.SWAGGER});
-
-      expect(spec).to.deep.equal({
-        definitions: {
-          Product: {
-            properties: {
-              title: {
-                type: "string"
-              }
-            },
-            type: "object"
-          }
-        },
-        tags: [
-          {
-            name: "Controller"
-          }
-        ],
-        paths: {
-          "/": {
-            post: {
-              operationId: "controllerMethod",
-              parameters: [],
-              produces: ["application/json"],
-              tags: ["Controller"],
-              responses: {
-                "200": {
-                  description: "description",
-                  schema: {
-                    properties: {
-                      data: {
-                        properties: {
-                          _id: {
-                            type: "string"
-                          },
-                          data: {
-                            $ref: "#/definitions/Product"
-                          }
-                        },
-                        type: "object"
-                      },
-                      totalCount: {
-                        type: "number"
-                      }
-                    },
-                    type: "object"
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-      expect(spec2).to.deep.equal({
-        definitions: {
-          Article: {
-            properties: {
-              id: {
-                type: "string"
-              }
-            },
-            type: "object"
-          }
-        },
-        tags: [
-          {
-            name: "Controller2"
-          }
-        ],
-        paths: {
-          "/": {
-            post: {
-              operationId: "controller2Method",
-              parameters: [],
-              produces: ["application/json"],
-              tags: ["Controller2"],
-              responses: {
-                "200": {
-                  description: "description",
-                  schema: {
-                    properties: {
-                      data: {
-                        properties: {
-                          _id: {
-                            type: "string"
-                          },
-                          data: {
-                            $ref: "#/definitions/Article"
-                          }
-                        },
-                        type: "object"
-                      },
-                      totalCount: {
-                        type: "number"
-                      }
-                    },
-                    type: "object"
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-    });
-    it("should declare a nested Generics of Model (openspec3)", async () => {
-      // WHEN
-      @Generics("T")
-      class Pagination<T> {
-        @CollectionOf("T")
-        data: T[];
-
-        @Property()
-        totalCount: number;
-      }
-
-      @Generics("T")
-      class Submission<T> {
-        @Property()
-        _id: string;
-
-        @Property("T")
-        data: T;
-      }
-
-      class Product {
-        @Property()
-        title: string;
-      }
-
-      class Controller {
-        @OperationPath("POST", "/")
-        @(Returns(200, Pagination).Of(Submission).Nested(Product).Description("description"))
-        async method(): Promise<Pagination<Submission<Product>> | null> {
-          return null;
-        }
-      }
-
-      // THEN
-      const spec = getSpec(Controller, {specType: SpecTypes.OPENAPI});
-
-      expect(spec).to.deep.equal({
-        components: {
-          schemas: {
-            Product: {
-              properties: {
-                title: {
-                  type: "string"
-                }
-              },
-              type: "object"
-            }
-          }
-        },
-        tags: [
-          {
-            name: "Controller"
-          }
-        ],
-        paths: {
-          "/": {
-            post: {
-              operationId: "controllerMethod",
-              parameters: [],
-              tags: ["Controller"],
-              responses: {
-                "200": {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        properties: {
-                          data: {
-                            items: {
-                              type: "object",
-                              properties: {
-                                _id: {
-                                  type: "string"
-                                },
-                                data: {
-                                  $ref: "#/components/schemas/Product"
-                                }
-                              }
-                            },
-                            type: "array"
-                          },
-                          totalCount: {
-                            type: "number"
-                          }
-                        },
-                        type: "object"
                       }
                     }
                   },
